+++ romancal/lib/suffix.py
@@ -53,6 +53,7 @@ SUFFIXES_TO_ADD = [
     "jump",
     "rampfit",
     "saturation",
+    "sourcedetection",
     "dark_current",
     "darkcurrent",
     "outlier_detection",
@@ -62,7 +63,7 @@ SUFFIXES_TO_ADD = [
 
 # Suffixes that are discovered but should not be considered.
 # Used by `find_suffixes` to remove undesired values it has found.
-SUFFIXES_TO_DISCARD = ["pipeline", "step"]
+SUFFIXES_TO_DISCARD = ["highlevelpipeline", "pipeline", "step"]
 
 
 # Calculated suffixes.
@@ -85,6 +86,7 @@ _calculated_suffixes = {
     "linearitystep",
     "dark_current",
     "jump",
+    "sourcedetection",
     "sourcedetectionstep",
     "tweakregstep",
     "outlierdetectionstep",