+++ romancal/regtest/test_linearity.py
@@ -9,7 +9,7 @@ from .regtestdata import compare_asdf
 @pytest.mark.bigdata
 def test_linearity_step(rtdata, ignore_asdf_paths):
     """Function to run and compare linearity correction files."""
-    input_file = "r0000101001001001001_01101_0001_WFI01_saturation.asdf"
+    input_file = "r0000101001001001001_01101_0001_WFI01_refpix.asdf"
     rtdata.get_data(f"WFI/image/{input_file}")
     rtdata.input = input_file
 
@@ -26,7 +26,7 @@ def test_linearity_step(rtdata, ignore_asdf_paths):
 def test_linearity_outfile_step(rtdata, ignore_asdf_paths):
     """Function to run and linearity correction files. Here the
     test is for renaming the output file."""
-    input_file = "r0000101001001001001_01101_0001_WFI01_saturation.asdf"
+    input_file = "r0000101001001001001_01101_0001_WFI01_refpix.asdf"
     rtdata.get_data(f"WFI/image/{input_file}")
     rtdata.input = input_file